producer/requirements.txt
kafka-python-ng==2.2.3
aerospike==16.0.0
